Fix some errors



See merge request !28

+++ app/controllers/home_controller.rb
@@ -2,7 +2,7 @@ class HomeController < ApplicationController
   
   def index
   	if current_user
-      @posts= Post.where("user_id IN(?) OR user_id = ?",Follow.select(:followable_id).where(:follower_id => 11), User.first.id).order("created_at DESC")
+      @posts= Post.where("user_id IN(?) OR user_id = ?",Follow.select(:followable_id).where(:follower_id => current_user.id), current_user.id).order("created_at DESC")
       render :feed
 	  end
   end
@@ -10,16 +10,17 @@ class HomeController < ApplicationController
   def add_post
     authenticate_user!
     post= Post.new(:content => params[:status])
-    current_user.post << post
+    current_user.posts << post
     redirect_to(:controller => 'home' ,:action => 'index')
   end
 
    def delete_post
     authenticate_user!
-    post= current_user.post.find(params[:id])
-    rescue ActiveRecord::RecordNotFound  
-     redirect_to :controller => "home", :action => "index"
-    return
+    post = current_user.posts.find_by(params[:id])
+    if post.nil? then
+     redirect_to(:controller => 'home' ,:action => 'index')
+     return
+    end
     post.destroy
     redirect_to(:controller => 'home' ,:action => 'index')
   end
